Bug #32545030 LONG SEMAPHORE IN SYNC0SHARDED_RW.H
The `lock_table_has_locks(const dict_table_t * table)` is documented as returning a potentially stale value, and the callers call it only in contexts where they expect the number of locks for a given table to be already zero or monotonically falling down towards zero.
There are two types of data-locks: LOCK_TABLE and LOCK_REC.
LOCK_RECs can be in any of the many shards of lock_sys, but are counted in the atomic `table->n_rec_locks`. This number can change at any moment unless you latch all lock-sys shards.
LOCK_TABLEs are all in the same shard for the given table.
The implementation in trunk took an expensive exclusive latch on whole lock-sys, to count all the locks in a correct way (that is: corresponding to a single particular moment in time), but this isn't really needed if the caller only cares if the number is already zero or not yet.
This patch latches only the shard containing the LOCK_TABLEs and accesses the n_rec_locks without any latch at all.
If the new implementation returns true, then it means that there was a moment in time when there was at least one data-lock, so the answer is as correct as in trunk.
If the new implementation returns false, then it only means that there was a moment in time then there was no LOCK_TABLE locks, and another moment when there was not LOCK_REC locks, but these could be two different moments. However, in the contexts this function is called we know the locks can not reappear, so the answer is also as correct as in trunk.
(Also, note that there is no way that I know of, which would let LOCK_REC to exist without there also being a LOCK_TABLE|LOCK_IS or LOCK_TABLE|LOCK_IX for the same table)

RB: 26003
Reviewed-by: Elżbieta Babij <elzbieta.babij@oracle.com>