Bug#32565479 - ROUTER FAILS TO START WHEN STATIC FIRST-AVAILABLE
Issue
=====

For static routing with first-available policy the listening sockets are
always open, which means that they are being opened at startup. Start
socket mechanism will fail if the router is not running yet. So there
is a race on startup between 'is running' and 'start socket acceptor'.

Fix
===

Check whether the routing plugin has stopped instead of checking if all
plugins are set up.

RB: 26009
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>