Bug#32541241: MANY THREADS IN OS_EVENT, SYSTEM SLOWDOWN/HANG
Problem
--------
The server slows down while attempting to open tablespaces as the current
number of open files is more than the variable innodb_open_files.
When a tablespace is opened, it is added to a LRU list whose size is defined
by the server variable innodb_open_files. When the number of files already
open reaches this limit and the server attempts to open another file, the oldest
file from this LRU list is closed.

The problem happens because even the temporary tablespaces are counted as
open files. These temporary tablespaces cannot be closed till the time
the user connections associated with them are active. This leads to a
situation where the number of opened files cannot be reduced further and hence
new files cannot be opened.

Solution
--------
Ignore temporary tablespaces while counting the open files.

RB: 26023
Reviewed By: Kevin Lewis <kevin.lewis@oracle.com>