Bug #32522398 ROUTER FAILS TO CONNECT ONLINE NODE1 WHILE CLOING NODE2
When calculating the quorum for the Cluster availability the Router counts Recovering
nodes in the quorum calculation. This however does not happen when the Recovering
node is missing in the metadata. This is the case in the cloning scenario: where
the node first appears in the GR tables and only after the cloning is finished is
also added to the metadata.

This patch modifies the logic so that the Recovering nodes that are not in the metadata
(only reported by GR) are not included in the quorum calculation (do not increase
the required pool).

RB: 26024