Bug#32579459 MEM LEAK WHEN CREATE EVENT OP FAIL
Memory leak occurs when the ndbcluster plugin fails to create a
subscription for receiving events from NDB. Ownership of the provided
event data is supposed to be taken over but that only happens when
create succeeds, in other cases the provided event data is lost.

Fix by moving the creating of Ndb_event_data into the same function that
creates the event operation, this makes it much easier to maintain the
lifecycle of the allocated Ndb_event_data. This change affects CREATE
TABLE, both copying and inplace ALTER TABLE as well as all the places
that setup binlogging of a table. All in all there are now fever error
cases.

Firm up the lifecycle of Ndb_event_data, create memory up front and
initialize most members in the constructor.

Improve and streamline management of the NdbEventOperation, its
Ndb_event_data and how it's installed and removed from NDB_SHARE::op and
the NdbApi. This is primarily done by always using same function to
"create or replace" the NdbEventOperation. There is also new function
for "installing" the NdbEventOperation into NDB_SHARE which makes sure
that the installed NdbEventOperation and its "custom data" is consistent
as well as required mutex and references are held.

Also fix a problem which might occur if the binlog thread restarts during
CREATE TABLE (and probably other DDL) where the NdbEventOperation was
allowed to be installed in a "dropped" NDB_SHARE. This is now handled
just as any other problem like above.

Change-Id: I3457f8e7546a866ec664f14465ac9d4cc1bd05ce

Bug#32579459 MEM LEAK WHEN CREATE EVENT OP FAIL

Review fix.

Forward declare ha_ndbcluster only once in ndb_share.h

Change-Id: I86f642a8673c4bc64e7457f030a78fa7a9dc66f7