BUG#32292413 GENERATED COLUMNS (VIRTUAL & STORED) DO NOT UPDATE ON REPLICA
Description
-----------
When using row-based-replication and while processing a `Write_rows_event`
or an `Update_rows_event`, stored generated columns that exist only on the
replica aren't being updated along the other fields in the row.

Analysis
--------
When processing a `Rows_log_event`, during the process of unpacking the row
images, stored generated columns that exist only on the replica aren't
being updated in the after-image and before the row being applied by the
storage engine.

Usually, we disregard the update of generated columns that also exist on
the source because: when they are virtual, the value will be calculated
upon each row read operation; when the column is stored, the stored value
will be included in the row image coming from the source (and, hence,
supporting non-deterministic functions as expressions for the stored
generated columns).

However, we should always updated stored generated columns that also exist
on the source if it's base column value has changed, since we don't know if
the expression that is used to generate the value didn't change.

In the case of the column only existing on the replica, we should evaluate
and update the column value in the row images.

Fix
---
In `Rows_log_event::unpack_current_row`, indentify stored generated columns
that need to be evaluated and, if any exist, update it's value in the table
row image about to be applied.

Reviewed-by: Neha Kumari <neha.n.kumari@oracle.com>
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
RB: 25890