Bug#32224707 CPU 100% MYSQL 8 ON WINDOWS 2019 - Regression
Issue
=====
As part of Bug#32224707 fix, we make the parallel reader infrastructure spawn
only as many threads as required, with innodb_parallel_reader_threads as the
max limit.

Initially we request for innodb_parallel_reader_threads number of threads by
calling Parallel_reader::available_threads. Once the actual number of threads
to be spawned is decided in Parallel_reader::run, we release the excess number
of threads and release the remaining m_n_threads as part of the destructor
function once the parallel read execution is done.

In RAPID's case though, it's possible that Parallel_reader::run is not called;
so m_n_threads is not set and we end up not releasing any of the threads that
we had requested for initially as m_n_threads will be 0. As a result, after a
few runs, the reader thinks that there are no available threads.

Fix
===
Make the RAPID adapters set m_n_threads as part of constructor instead of doing
it in Parallel_reader::run.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>
RB: 25899