Bug #32618829 USE MERGE_LIBRARIES_SHARED FOR THE NDBCLUSTER CLIENT LIBRARY.
The MySQL client library libmysqlclient.so is created with the cmake
macro MERGE_LIBRARIES_SHARED, which merges a set of known static
libraries, and applies various platform-dependent compile/link
commands to make a valid .so/.dll library which exports a desired set
of symbols.

This patch uses the same approach for creating libndbclient.so, the
ndbcluster client library.

The goal is to simplify the cmake code, and improve maintainability.
The next step would be to get rid of the various uses of the cmake
cache entries for <tgt>_LIB_DEPENDS, and maintain library dependencies
ourselves (and get rid of the CMAKE_POLICY(SET CMP0073 OLD)).

The patch cleans up the Windows and Xcode builds by distinguishing
between ADD_DEFINITIONS, which applies to all build types, and
CMAKE_CXX_FLAGS_<config> which can be targeted at individual build
types. We remove the setting of compiler flag -DNDEBUG, which belongs
in CMAKE_CXX_FLAGS_RELWITHDEBINFO (and others) rather than as
ADD_DEFINITIONS.

For the Windows build, we use ADD_SHARED_LIBRARY, rather than
MERGE_LIBRARIES_SHARED for creating the ndbclient.dll. The reason for
this is that the library exports a large set of mangled C++ names,
rather than a fixed set of symbols.

The patch also uncovered a bug in the existing linkage of
libndbclient. All ndbcluster test executables were linked incorrectly,
because of this:
  TARGET_LINK_LIBRARIES(ndbclient_so ${NDBCLIENT_SO_LIBS})
The net result of that line was that all test clients were linked with
all the static libraries, in addition to libndbclient.so

Change-Id: Idbc5749c3f9081a90cfc05401ecf1c0f0b15942e