Bug#32611913: ASSERT FAILURE IN ITEM_FUNC_MATCH::EQ()
Item_func_match::eq() raised an assert failure in debug builds if it
was called with an argument that was an Item_func_match_predicate.
The assertion was added because it was believed that an
Item_func_match object would not be compared to an
Item_func_match_predicate object. This can however happen in the
ONLY_FULL_GROUP_BY check if the predicate is in a HAVING clause.

Fixed by removing the assertion so that Item_func_match::eq()
completes and returns false.

Change-Id: I8be56ad89011155731cf7bd7df7ec0aee8133db6