Bug#32617433: SIGSEGV IN HYPERGRAPH OPTIMIZER WITH INSERT SELECT ODKU
The server could exit when using the hypergraph optimizer for an
INSERT SELECT statement with an ON DUPLICATE KEY UPDATE clause, and
the source was an aggregated query.

The problem occurs when the optimizer tries to update the ODKU clause
by replacing references to columns in the select list with the
corresponding columns in a temporary table. If there are hidden
fields, it ends up trying to replace select list references with
non-corresponding temporary table columns.

Fixed by looking at the visible fields only.

Change-Id: I575d0ae4748c19494b3ecd5e2f3156c7108f0d76