Bug#32622880 INCONSISTENT USE OF LOGPAGEARRAY IN DBLQH::READEXECLOG
PROBLEM
=======

In Dblqh::readExecLog there are early checks that the number of pages to
read is at most 15, while when composing the GSN_FSREADREQ signal
reading more than 9 pages will result in an incomplete signal.

As long as an prepare operation record is less than one page one would
not need read more than two pages if record cross a page boundary.

That should make this bug harmless causing no actual malfunction.

Code is inconsistent and misleading and should be corrected.

FIX
===

The logPageArray in Dblqh::LogFileOperationRecord will be adjusted to 9
pages to reflect the worst case supported by old code.

Also the storage of the file offset in logPageArray is removed, it is
already available from several places.

Change-Id: I8d86a68b7f9f820edceb092dcd17890b6ae60ed5