BUG#32553818 ENFORCE USING --INITIAL OR --RELOAD WHEN STARTING MGMD WITH -F OPTION
When mgmd is given -f option, it should be always started with
--initial or --reload options. Without these options mgmd should not
start and throw an error to the user.

This patch adds a validation that checks if --initial or --reload
option is specified along with -f/config-file and if this check fails
mgmd is not started and error message is displayed to users.

Change-Id: Icdebb3b3cfa53bc26f9da45d9286de0d2f5def8a