Bug #32530147 	ASSERTION FAILURE: BUF0FLU.CC:1163:BPAGE->IS_DIRTY() THREAD 139641244079872
Fixed `buf_flush_ready_for_flush()` to not claim page as ready for flush when it is stale. We need first check if it is dirty and not IO fixed. After this check, we will return true either way, if it is possible, so there is no need for the `if (bpage->was_stale()) { return true; }`.
Unused `buf_phy_read_ahead()` is removed.
`buf_block_hash_get_locked` apparently was never used with `lock == nullptr`, but potentially it could crash.

RB#26050