Bug#29634540 "DROP DATABASE OF 1 MILLION TABLES RESULTED CRASH OF MYSQLD" [noclose]
Attempt to drop database with 1 million tables results in
out-of-memory condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of the problem was that DROP DATABASE for each update of
data-dictionary it did allocated temporary proxy LEX object on
statement memory root. As result even though LEX object was needed
for duration of individual update operation only its memory was
not reclaimed until statement's end.

This patch solves the above issue by changing code responsible for
data-dictionary updates to use, instead of memory-root-allocated proxy
LEX objects, smaller Query_tables_list objects which are allocated
on heap and deleted immediately after each update.

There is no MTR test case as it is hard to write robust one for
this situation.

This change doesn't solve the original problem to the full extent
but is one in a series of patches which aim to do this.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldvoll@oracle.com>