Bug#32642657: FAILURE WHEN REEXECUTING INSERT SELECT WITH HYPERGRAPH
Some INSERT SELECT prepared statements with an ON DUPLICATE KEY UPDATE
clause fail on the second execution when the hypergraph optimizer is
used. The problem happens if the SELECT statement accesses the target
table (so that SQL_BUFFER_RESULT is implied) and the ON DUPLICATE KEY
UPDATE clause sets a column in the target table to the value of a
column in the source table. The ON DUPLICATE KEY UPDATE clause is
modified in the first execution with data that gets freed when the
execution ends. When this data is accessed in the next execution, the
server crashes while trying to access the freed data.

Fixed by using THD::change_item_tree() when updating the ON DUPLICATE
KEY UPDATE clause, which makes sure the changes are rolled back at the
end of the execution.

Change-Id: I1c9faa4351ab425bdd72d806c97223ba8f984164