Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [simplify order/distinct, noclose]
Make the logic in ApplyDistinctAndOrder() much simpler; we decide
beforehand whether we need to sort by row IDs or not, so that we don't
need to create the filesort objects early (which would otherwise
give us that answer).

This is possible largely because WL #14325 removed the main edge case of
sorting by constant expressions only, which would earlier require row
IDs. It is also greatly helped by the new and more generic CollectTables()
introduced in the previous patch, relieving us from having to track the
table list ourselves as we add temporary tables.

In theory, we could be slightly more aggressive about avoiding row ID
sorts; e.g., if we have a materialization at some point that removes all
remaining blobs (perhaps they're only used in some expression), we don't
really need row IDs for sortahead sorts, but this is marginal. So at
this point, we enshrine fairly early on that all sorts are either on row
IDs or they are not.

Change-Id: I9ef89ad4cbdf943de2eb7b5e8eeb06262f29c80b