Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [remove copy_fields, noclose]
When copying items into a temporary table row, we have different
handling based on whether the Item is an Item_field or not; we go and
look at the underlying Field, and do special handling of it. This causes
a bifurcated copy system, where everything needs to consider both
param->copy_fields and param->items_to_copy, and the identity of the
Item_field gets lost if one wishes to track Items through temporary
tables.

There's no good reason to do this, so just handle everything as
items_to_copy. The only thing copy_fields is used for now, is to copy
fields back and forth between window function's framebuffer tables;
it is needed there because it can also reverse the direction of copying.

Note that this change causes slightly different handling of copying of
timestamps around timezone handling (since that happened to be different
between Item_field and Field), making it both slightly more consistent and
actually also often more correct (e.g., time_zone.test no longer gets an
error, but correctly keeps the two timestamps apart -- this is mostly
luck, though, and will still depend on the chosen query plan).

Change-Id: Ic2824366822bbf7718ceed945155ec130910aa4b