WL#14422: Full-text search in the hypergraph join optimizer [patch 10/12, limit avoid sort]
Avoid more sorts for ORDER BY on descending score by using an ordered
full-text index scan. If the query accesses a single table, and it
does not use aggregation or DISTINCT, it is sometimes possible to use
a full-text index scan even if there is no WHERE clause that ensures
that only matches are returned. This can be done if the query has a
LIMIT which is lower than the number of matches found.

This optimization is for InnoDB only, since MyISAM doesn't support
retrieving the number of matching rows from the full-text index.

Change-Id: If8eb6ea3ee51b3f280786138cf1855d340140cff