BUG#32655364: MEMORY LEAK IN SOURCE_IO_MONITOR
Source_IO_monitor has a static shared instance that was not being
properly terminated. The root issue was the instance being static,
suffering from out of order initialization and destruction, in
particular the external performance schema instrumentation global
references it pointed to were cleaned before it.

To solve the above issue, now this instance is a singleton on
mysqld.cc with clear defined initialization and destruction points.

RB: 26107