Bug#32173664 - INSERT WRONG TIMESTAMPS USING EXPLICIT TIME ZONE OFFSET IN TIME_ZONE WITH DST
To check for boundaries of a given timestamp with timezone
offset, the value is converted into a timestamp value without
timezone and in the process a call is made localtime_r(). The
value returned is used for further processing and storing in the
field.

This timestamp conversion always returns the higher hour
for the repeat hour during end of DST and so wrong values gets
stored in the field.

Fix:
Use the converted timestamp for bounds checking. Continue to use
the timestamp with timezone offset value for storing into the
field.

Change-Id: Iebc16597899f5e7fba105e526df3b7f1ae6feae3