Bug#32612855: LOG_SINK_PFS_EVENT_ADD: ASSERTION `E->M_MESSAGE_LENGTH > 0' FAILED.
Under very specific circumstances, reading a previous run's error log
can produce log-events for performance_schema.error_log that do not
contain a message (but have the other fields). This is largely harmless,
but of interest, and therefore triggers the assertion in the title.
Since this issue has been triggered in pushbuild runs, this patch
adds (slightly) better diagnostics. This should make it easier to
identify such events where desired without needing an assertion that
pollutes the pushbuild run. The server will still assert() if such
an event is generated elsewhere in the code; this is intentional.

Change-Id: I8b43a7c782fd872098569fd4d3c44991cf744cc6