Bug#32672577 ALLOW THE METADATA_CACHE TO USE THE CLUSTER IDENTIFIED BY UUID OR ROLE.
In the current implementation during the bootstrap the Cluster name is written
to the static configuration file and then the metadata is queried for the
Cluster with a given name.
This is limiting and somewhat wrong as the cluster name can be changed
using Admin API (this bug is not for that particular problem).
Options to follow the cluster should be more flexible:
it should be possible to follow it by the UUID or it's role in the ClusterSet
in the future (use the Cluster that is the current primary).

This patch refactors the code in a way that makes it possible to follow
the Cluster also by it's UUID or role in the ClusterSet.

RB: 26132