Bug #32715381 UT0NEW ALLOCATOR CONSTRUCT() MAKES UNNECESSARY COPIES
ut_allocator implements construct and destroy interfaces in pre-C++11
style which do not take advantage of perfect-forwarding. Therefore,
any client or library code that is using ut_allocator to construct new
instances of T is going to have a larger cost than needed. For example,
  (1) Caller or library code first must instantiate a _temporary_
      instance of T.
  (2) Caller or library code must pass the reference to that temporary
      to construct(pointer p, const_reference val).
  (3) ut_allocator::construct then must use that temporary instance of
      T to _copy-construct_ the new and wanted instance of T.
  (4) Compiler must generate code to destroy the temporary instance.

Starting with C++11, none of these steps are anymore required and can be
replaced with more optimal, in terms of CPU and resources, alternative of
perfect-forwarding that supports in-place construction. This means that
no temporaries are going to be created, no references are going to be
involved, no copy-construction is going to take place anymore and finally
no temporaries are going to be needed to be destroyed. With perfect-forwarding,
all of these steps will be replaced by a single move-construction.

In the context of repercussions to an actual code, this means that only
because ut_allocator provides a non-perfect-forwarding variant of
construct interface, all STL containers that are declared to be used
with ut_allocator are going to be doing all of the above steps silently.
This is of course suboptimal in terms of performance.

Fixing this situation can either be done by providing the implementation
of perfect-forwarding ut_allocator::construct variant or it can also be
done by completely cutting loose ut_allocator from having _any_
variant of construct implementation.

That means that in absence of construct/destroy in custom allocator
implementation, any conforming STL container implementation must first
look for existence of std::allocator_traits<my_custom_allocator>
specialization, and if it doesn't find any it must fallback to default
non-specialized implementation provided by std::allocator_traits<Alloc>.

It happens that std::allocator_traits<Alloc> default implementation is
exactly a perfect-forwarding approach so this patch fixes this unnecessary
overhead by removing the ut_allocator::construct implementation
completely. Same would have been done with ut_allocator::destroy but
some recent changes to Page_load code made it temporarily impossible.

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26271