Bug #32715359 UT0NEW ALLOCATOR DOUBTFUL OOM HANDLING
ut_allocator in the event of OOM is tasked to report a message through
InnoDB logging infrastructure but this is highly doubtful.

This is because there's hardly much what can be done in cases when, and
if (*), malloc returns a nullptr (OOM). Even if the code wants to do
something in OOM cases then such code must be very carefully crafted with
many things in mind but at least it must be such that it does not require
any subsequent dynamic allocations.

ut_allocator is doing exactly that. It is trying to report a message in
the event of OOM but on a closer look one can see that crafting the
message itself requires more dynamic allocations ([1], [2]). That of
course doesn't serve the purpose and therefore this operation of reporting
the message in the event of OOM is deemed to fail from the beginning.

Given that the OOM reporting mechanism cannot live to its promise, and
has probably never been tested against the OOM scenario, this patch is
removing it. FWIW ut_allocator will become less statefull in PFS builds
and stateless in non-PFS builds.

[1] ib::fatal_or_error is implemented in terms of std::string and
    std::ostringstream. Both of them dynamically allocate memory.
    See storage/innobase/include/ut0ut.h
[2] strerror is allowed to use malloc() which of course also ends up
    doing more dynamic memory allocations.
    See https://code.woboq.org/userspace/glibc/string/strerror.c.html#strerror

(*) Detecting that malloc returned a nullptr on systems which are
configured to run in over-commit mode (which is a default) is not even
possible simply because malloc on such systems never returns a nullptr.
That makes this OOM reporting mechanism not only arguable in theory as
it has been described above but also in practice.

malloc on such systems will only ever create a record that there was an
allocation request but it will not materialize the request by actually
consuming the real physical memory. Therefore it is designed as such that
it will _always_ return a valid pointer but the application code will not
know whether or not there is enough underlying physical memory to
materialize the allocation request until the very moment when application
code tries to access the memory through that pointer.

At that moment, page-fault is going to be generated as there will be
existing no corresponding TLB entry at that point and depending on if
there are enough resources available (physical RAM, swap space left on
secondary storage), accessing that memory will either be successful
(but preceded by either minor or major page-fault) or will follow with
the application crash because OS did not manage to find any free space
after complex major page-fault handling algorithms. That is the moment
when it's already too late to be able to do anything and it makes the
situation in code impossible to recover from.

Only somewhat reasonable part in InnoDB code that may employ some form
of so called "OOM handling" is the buffer-pool initialization code.
MySQL during the startup runs the InnoDB initialization which in turn
runs the buffer-pool initialization code whose purpose is to pre-allocate
unusually large chunks of memory (by default each chunk is 128MB large
but it is configurable).
In case that:
  (1) system is _not_ running in over-commit mode (e.g. Windows) _and_
  (2) system really ran out of the available memory to service those
      large allocation requests,
MySQL could during its startup, other than just throwing std::bad_alloc
which is already self-explanatory and what is happening with this patch,
put some more extra information into the log about the failed allocation.
There's of course also no guarantee that logging operation will succeed
but given the special circumstances and knowing that large allocation is
what has failed, then it's more likely that there is still enough bytes
left to carry on with the operation. Code at this point can even perhaps
implement try-harder retry-loop logic in order to possibly recover from
this situation. Clearing up the code around these places is going to be
addressed in soon to come patches.

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26273