Bug#29634540 "DROP DATABASE OF 1 MILLION TABLES RESULTED CRASH OF MYSQLD" [noclose]
Attempt to drop database with 1 million tables results in
out-of-memory condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of the problem was that DROP DATABASE for each table, tablespace,
routine or event it has dropped allocated full-blown copy of respective
DD object and stored it in thread's Dictionary_client registry of
dropped objects. These objects were kept in this registry until
statement's end. Since each individual object contained copy of
all sub-objects of the original object (even though this information
as irrelevant for tracking dropped objects) they hogged quite some
memory in total (in my model test that has dropped database with
5K of tables it was responsible for around 25% of memory consumed
by server).

This patch addresses the issue by replacing use of full-blown copies
of DD objects in thread's dropped objects registry with special
placeholders. These placeholders are almost-empty versions of
original objects with only their key members set, which is the only
info relevant for dropped objects registry.

This change doesn't solve the original problem to the full extent
but is a big step in this direction.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldvoll@oracle.com>