Bug #32618301 INNODB SHOULD REPORT CHANGES IN WAIT-FOR GRAPH TO SERVER LAYER
InnoDB's lock system only informed server about the first t2 -waits-for-> t1 edge which caused t2 to go to sleep, but when t1 released the lock and t2 still had to wait for another transaction, say t2 -waits-for-> t3, InnoDB did not inform the server about this.
This could impact replication, as coordinator was not aware that t2 waits for t3, and thus could not detect a conflict with the intended commit-oder (t1, t2, t3).

RB: 26060
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>