WL#14422: Full-text search in the hypergraph join optimizer [patch 11/12, covering index]
Enable use of covering index scans for full-text search when the query
only needs to access the document ID and the rank. Covering scans are
currently only used if the storage engine supports the extended
full-text API and the table has an explicit FTS_DOC_ID column (these
limitations also apply to the old optimizer).

Change-Id: I0bc1814185c49ebffa0621ae7099ef652767dbd1