BUG#32666435: DUMP FILE IS NOT DELETED AFTER INCREMENTAL BACKUP USING PAGE               TRACKING IN DATADIR
Description:
------------
When page-track is enabled during an incremental backup, MEB runs a UDF to get
a list of changed pages. Running this UDF creates a file in the server data
directory which contains information about the changed pages and the path to
this file is stored in a static variable. After the UDF finishes, the memory
allocated for this static variable is de-allocated and the value is set to
nullptr.

At the end of a backup operation, MEB calls the UDF to update the backup ID to
NULL, which also takes care of removing the file created in the server data
directory.

However, when this UDF is called, the value of the static variable holding the
path value is nullptr and the component couldn't delete the file.

Fix:
----
Don't free the static variable holding the path value for the page-track file
after the UDF to get the list of changed pages is finished. Instead, at the
end of backup, when MEB calls the UDF to update the backup ID to NULL, free
the static variable and set its value to nullptr after removing the page-track
file created in the server data directory.

RB: 26150