Bug#29634540 "DROP DATABASE OF 1 MILLION TABLES RESULTED CRASH OF MYSQLD" [noclose]
Attempt to drop database with 1 million tables results in
out-of-memory condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of the problem was that both code which acquired metadata locks
on tablespaces for tables to be dropped and InnoDB code retrieving
information about tablespaces while dropping tables created an uncached
dd::Tablespace object for each table/tablespace, which were freed only
at statement's end, even though there is no real need to keep them
for that long.

This patch solves this issue by changing code using those uncached
dd::Tablespace objects to release them as soon as they become
unnecessary.

There is no MTR test case as it is hard to write robust one for
this situation.

This change doesn't solve the original problem to the full extent
but is yet another step on the road to this.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldoll@oracle.com>