Bug #32688540: MATERIALIZE ITERATORS HAVE SKEWED TIMINGS FROM SKIPPED MATERIALIZATIONS
On EXPLAIN ANALYZE, Materialize iterators count every single Init()
call, even those that just keep the existing data. This is confusing and
a bit misleading, since materializations then appear to be way too cheap
compared to the number of calls underneath them. Fix by allowing
MaterializeIterator to override the call and row counts with its own
data. This skews the timings ever so slightly the other way (we count
the time used to figure out we didn't need a materialization), but it
should generally be within the noise.

Other materializing iterators (like table functions) may have similar
requirements that we may wish to fix at some point, but we're leaving
this for future improvement.

Change-Id: I17e7f6d4e26f80ac3618c1b034f69e3a7a4936cc