Bug#32808324 SIGHUP ignored right after sd-notify READY
Issue
=====

routertest_component_logging fails for SIGHUP related tests on slow
machines as simulated with:

  $ taskset -c 0 ./routertest_component_logging

The test waits for router to start, signal READY via sd_notify() and
then sends a SIGHUP to trigger a log-file rotation:

1. wait for sd-notify READY
2. send SIGHUP to router

Router's steps:

1. set signal handler for SIGHUP:
   if (g_log_reopen) g_log_reopen();
2. signal READY via sd-notify
3. set global g_log_reopen

If the signal is received between 2. and 3. it will get ignored and
logfile will not be rotated ... and the test will fail.

Change
======

- signal READY after g_log_reopen is set
Flip the steps for the router:

1. set signal handler for SIGHUP:
   if (g_log_reopen) g_log_reopen();
2. set global g_log_reopen_thread
3. signal READY via sd-notify

- use Monitor<> to synchronize access to g_log_reopen

RB: 26285